# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de Cobrança Mensal por Cidades
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface City {
  id: number;
  name: string;
  /**
   * @minimum 1
   * @maximum 28
   */
  dueDay: number;
  notes?: string | null;
  createdAt: string;
}

export interface CityWithSummary {
  id: number;
  name: string;
  dueDay: number;
  notes?: string | null;
  payersCount: number;
  expectedTotal: number;
  receivedTotal: number;
  pendingTotal: number;
  overdueTotal: number;
}

export type PayerWithStatusStatus =
  (typeof PayerWithStatusStatus)[keyof typeof PayerWithStatusStatus];

export const PayerWithStatusStatus = {
  paid: "paid",
  pending: "pending",
  overdue: "overdue",
} as const;

export interface PayerWithStatus {
  id: number;
  cityId: number;
  name: string;
  monthlyAmount: number;
  contact?: string | null;
  notes?: string | null;
  status: PayerWithStatusStatus;
  paymentId?: number | null;
  paidAt?: string | null;
  paidAmount?: number | null;
}

export interface CityDetail {
  city: CityWithSummary;
  referenceMonth: string;
  payers: PayerWithStatus[];
}

export interface CreateCityBody {
  /** @minLength 1 */
  name: string;
  /**
   * @minimum 1
   * @maximum 28
   */
  dueDay: number;
  notes?: string | null;
}

export interface UpdateCityBody {
  /** @minLength 1 */
  name?: string;
  /**
   * @minimum 1
   * @maximum 28
   */
  dueDay?: number;
  notes?: string | null;
}

export interface Payer {
  id: number;
  cityId: number;
  name: string;
  monthlyAmount: number;
  contact?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreatePayerBody {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  monthlyAmount: number;
  contact?: string | null;
  notes?: string | null;
}

export interface UpdatePayerBody {
  /** @minLength 1 */
  name?: string;
  /** @minimum 0 */
  monthlyAmount?: number;
  contact?: string | null;
  notes?: string | null;
}

export interface Payment {
  id: number;
  payerId: number;
  /** Formato YYYY-MM */
  referenceMonth: string;
  amount: number;
  paidAt: string;
  notes?: string | null;
}

export interface RegisterPaymentBody {
  /** @pattern ^\d{4}-\d{2}$ */
  referenceMonth: string;
  /** @minimum 0 */
  amount: number;
  paidAt?: string | null;
  notes?: string | null;
}

export interface DashboardSummary {
  referenceMonth: string;
  citiesCount: number;
  payersCount: number;
  expectedTotal: number;
  receivedTotal: number;
  pendingTotal: number;
  overdueTotal: number;
  paidPayers: number;
  pendingPayers: number;
  overduePayers: number;
}

export interface ActivityEntry {
  paymentId: number;
  payerId: number;
  payerName: string;
  cityId: number;
  cityName: string;
  amount: number;
  referenceMonth: string;
  paidAt: string;
}

export type GetDashboardParams = {
  /**
   * Mês de referência no formato YYYY-MM. Padrão é o mês atual.
   * @pattern ^\d{4}-\d{2}$
   */
  referenceMonth?: string;
};

export type ListCitiesParams = {
  /**
   * @pattern ^\d{4}-\d{2}$
   */
  referenceMonth?: string;
};

export type GetCityParams = {
  /**
   * @pattern ^\d{4}-\d{2}$
   */
  referenceMonth?: string;
};

export type ListRecentActivityParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};