# lib/api-zod/src/generated/types/activityEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de Cobrança Mensal por Cidades
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEntry {
  paymentId: number;
  payerId: number;
  payerName: string;
  cityId: number;
  cityName: string;
  amount: number;
  referenceMonth: string;
  paidAt: Date;
}