# lib/api-zod/src/generated/types/cityWithSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de Cobrança Mensal por Cidades
 * OpenAPI spec version: 0.1.0
 */

export interface CityWithSummary {
  id: number;
  name: string;
  dueDay: number;
  notes?: string | null;
  payersCount: number;
  expectedTotal: number;
  receivedTotal: number;
  pendingTotal: number;
  overdueTotal: number;
}