# lib/api-zod/src/generated/types/createCityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de Cobrança Mensal por Cidades
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCityBody {
  /** @minLength 1 */
  name: string;
  /**
   * @minimum 1
   * @maximum 28
   */
  dueDay: number;
  notes?: string | null;
}