# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de Cobrança Mensal por Cidades
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  referenceMonth: string;
  citiesCount: number;
  payersCount: number;
  expectedTotal: number;
  receivedTotal: number;
  pendingTotal: number;
  overdueTotal: number;
  paidPayers: number;
  pendingPayers: number;
  overduePayers: number;
}