# lib/api-zod/src/generated/types/payer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de Cobrança Mensal por Cidades
 * OpenAPI spec version: 0.1.0
 */

export interface Payer {
  id: number;
  cityId: number;
  name: string;
  monthlyAmount: number;
  contact?: string | null;
  notes?: string | null;
  createdAt: Date;
}