# lib/api-zod/src/generated/types/payerWithStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de Cobrança Mensal por Cidades
 * OpenAPI spec version: 0.1.0
 */
import type { PayerWithStatusStatus } from "./payerWithStatusStatus";

export interface PayerWithStatus {
  id: number;
  cityId: number;
  name: string;
  monthlyAmount: number;
  contact?: string | null;
  notes?: string | null;
  status: PayerWithStatusStatus;
  paymentId?: number | null;
  paidAt?: Date | null;
  paidAmount?: number | null;
}