# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de Cobrança Mensal por Cidades
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: number;
  payerId: number;
  /** Formato YYYY-MM */
  referenceMonth: string;
  amount: number;
  paidAt: Date;
  notes?: string | null;
}